+++ app/models/central-worship-service.js
@@ -1,8 +1,3 @@
 import WorshipAdministrativeUnitModel from './worship-administrative-unit';
-import { CentralWorshipServiceCodeList } from '../constants/Classification';
 
-export default class CentralWorshipServiceModel extends WorshipAdministrativeUnitModel {
-  get isCentralWorshipService() {
-    return this._hasClassificationId(CentralWorshipServiceCodeList);
-  }
-}
+export default class CentralWorshipServiceModel extends WorshipAdministrativeUnitModel {}
